Kinematic Equations from paper
The closed-form solution Kinematic Equations for "Exact Motion Planning for Tractor Trailer Robots" paper. 
One can manipulate the tractor velocity and steering angle to generate a trajectory. 
It needs to include the Rotational Path and Translational Path logic as outlined in Section 3.1.1 and 3.1.2 respectively. 
I expect the motion planner to come up with strategies based on a series of Paths, stretches, and bends.