5th Wheel 2nd DOF and Physics
Added the 2nd DOF to the 5th wheel since the trailer was being lifted into the air by the heavier cab pivoting on the rear wheels of the cab.  Reduced the mass by a factor of 100 and increased wheel friction by a factor of 4 to improve steering dynamics.  Steering still "bumps" periodically and using small steering angles causes no change in direction.